requirements.txt
# ----
# This modules should already be installed by default by installing the packages above

Flask==2.1.0
# itsdangerous
# Jinja2
# MarkupSafe
# Werkzeug

# ---
# Gunicorn module

gunicorn==20.1.0

# ---
# Base python packages

apiflask
flask-cors
boto3
google-cloud-bigquery

# ---
# Scheduler related

apscheduler
requests_futures

# ---
# Ultralytics packages (Not installed by default)

ultralytics
py-cpuinfo
# lapx>=0.5.2
# pafy
# youtube_dl